#42: Обработка ошибок в методе сервиса
